Make host a submodule
